+++ Web/SFB.Web.UI/Models/SchoolViewModel.cs
@@ -156,7 +156,7 @@ namespace SFB.Web.UI.Models
             }
         }
 
-        public int TotalPupils => ContextDataModel.NumberOfPupils.GetValueOrDefault();
+        public float TotalPupils => ContextDataModel.NumberOfPupils.GetValueOrDefault();
 
         public string IsPost16 => ContextDataModel.OfficialSixthForm == "Has a sixth form" ? "Yes" : "No";
